Total project structure changes

+++ Lesson3/src/circles.hpp
@@ -74,7 +74,7 @@ public:
         }
     }
 private:
-    int minSize;
+    size_t minSize;
     VC& circles;
 
 };
